refactor test code and src
