+++ contracts/test/EControllerTest.sol
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: MIT
+pragma solidity 0.8.2;
+
+import "../EController.sol";
+
+contract EControllerTest is EController {
+    constructor()
+    {
+        initialize();
+    }
+}
